effect startPos / endPos as coord3D and use to make subsets of physleds
general
- use UINT8_MAX and UINT16_MAX instead of own defines

index.css
- canvas has black border
- table has box-shadow (as border-radius alone does not work)

index.js
- coord3D: add change event to x,y,z input nodes
- sendvalue: add "value" property, add support for coord3D

AppLedsV: fixtureProjectAndMap:
- default y and z to 0 (not 1)
- only process if physical led within startPos and endPos boundaries
- 2D-2D: use start and endpos to calc virtual coordinate and bucket
- other dimensions: to do! (WIP)

AppModLeds:
- fxStart and End chFun: assign ledsV.start/endPos
- move dimensions and nrOfLeds to table

SysModUI: processJson:
- add vector to freeze for loop
- value handling: refactor and allow for coord3D, add setValue<JsonObject> for coord3D

SysModWeb
- wsEvent: add 💀 for multipart messages (WIP)

+++ src/App/AppModFixtureGen.h
@@ -726,11 +726,12 @@ public:
       
       genFix.openHeader("3DSideCube%d%d", length, sides);
 
+      //use (uint16_t)(length+1) to surpress warning warning: narrowing conversion of '(((int)length) + 1)' from 'int' to 'uint16_t' {aka 'short unsigned int'} inside { } [-Wnarrowing]
       genFix.plane3D({1, 1, 0},   {length, length, 0}, true); // front (z=0, first x then y)
-      genFix.plane3D({length, length+1, length}, {1, length+1, 1}, false); // bottom (y=length+1, first x, then z)
-      genFix.plane3D({length+1, 1, 1}, {length+1, length, length}, false); // right (x=length + 1, first z, then y)
+      genFix.plane3D({length, (uint16_t)(length+1), length}, {1, (uint16_t)(length+1), 1}, false); // bottom (y=length+1, first x, then z)
+      genFix.plane3D({(uint16_t)(length+1), 1, 1}, {(uint16_t)(length+1), length, length}, false); // right (x=length + 1, first z, then y)
       genFix.plane3D({0, length, length}, {0, 1, 1}, true); // left (x=0, first y, then z)
-      genFix.plane3D({length, 1, length+1}, {1, length, length+1}, true); // back (z = length+1, first x, then y)
+      genFix.plane3D({length, 1, (uint16_t)(length+1)}, {1, length, (uint16_t)(length+1)}, true); // back (z = length+1, first x, then y)
       genFix.plane3D({1, 0, length}, {length, 0, 1}, true); // top (y=0, first x, then z)
 
       // genFix.sideCube3D (0, 0, 0, length, sides);
